requirements.txt
gradio 
openai 
gtts 
pydub 
numpy 
requests 
whisper 
groq 
python-dotenv